==> 00_analysis_instMapping.ipynb <==
# First analysis and instruments aggregation

This notebook starts from the file 'HOP_OP_Tradeteq_09272018_', which is generated by the script 'extract_data.py' which translates columns names from the original file (they were not in English), converts all the dates to pandas datetime, converts all the numeric strings in numeric values (float) and deals with some typo found on the data provided.   
In this notebook a first data exploration for instruments aggregation is performed.

## Data import

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import datetime

#Receivables analysis
#import data
filename = "00_transactionsdf_simNames.pkl"
datafolder = "../.."+"/data/"
ReportDate = datetime.datetime(2018, 9, 28) #reference date on which data are received - all due dates after report date are related to open instruments

tr = pd.read_pickle(datafolder+filename)

tr.head()

tr.shape

tr.describe()

#should be datetime64[ns] for all date columns
tr.dtypes

tr.columns = tr.columns.str.strip().str.lower().str.replace(' ', '_').str.replace('(', '').str.replace(')', '')

tr.head().transpose()

#different values for string type columns in the dataset
print(tr['discharge_type'].unique(), len(tr['discharge_type'].unique()))
print(tr['transaction_type'].unique(), len(tr['transaction_type'].unique()))
print(tr['booking_text'].unique(), len(tr['booking_text'].unique()))
print(tr['factoring_type'].unique(), len(tr['factoring_type'].unique()))
print(tr['test_feature_tested'].unique(), len(tr['test_feature_tested'].unique()))
print(tr['currency'].unique(), len(tr['currency'].unique()))
print(tr['document_status'].unique(), len(tr['document_status'].unique()))

The column 'document_status' (present in the previous dataset but not in this one) had 4 different values, that can be translated in the following way:  
- 'historisch' = historical data
- 'storniert' = canceled
- 'offen' = open
- 'ungebucht' = unaccounted

## Instruments mapping

tr.shape

### 1. Relabel invoices to ensure unique ids

At this step, the script will perform an aggregation of unique ids at instrument level.  
Each transaction is part of one and only one unique instrument which can be isolated using the following fields:  
- invoice number
- debtor_id
- bill_id  

We can create a new column called 'uid' containing the unique id of each instrument.

#key_columns = ['invoice_number',  "debtor_id", "bill_id"] 
tr["uid"] = tr['invoice_number'].map(str) + ":" +tr["debtor_id"].map(str)+"/" + tr["bill_id"].map(str)

#drop columns with identical values
to_drop = []
for c in tr.columns:
    if len(tr[c].unique())==1:
        print(c)
        to_drop.append(c)

tr = tr.drop(columns = to_drop)
tr.shape

tr.head(10).transpose()

### 2. Replace transaction types with a numeric dictionary

Transaction types that can be associated with an entering flow of money (to be combined with the 'paid_amount' field):
- 'incoming payments'
- 'Credit entry'
- 'Subsequent assignment of the payment'
- 'Cancel invoice client'  
  
Transaction types that can be considered as an exiting flow of money (to be combined with the 'purchase_amount' field):
- 'Purchase invoice'
- 'Additional purchase invoice'

tr.drop(tr[tr.transaction_type.isnull()].index, axis=0, inplace=True)

ttypes = ['Purchase invoice', 'incoming payments', 'Reminder bill',
                      'Additional purchase invoice', 'Bulk confirmation',
                      'Subsequent assignment of the payment',
                      'Cancel invoice client',
                      'Cancellation invoice',
                      'Debt collection invoice',
                      'Credit entry', 'Cancellation of the payment',
                      'Subsequent allocation of the credit',
                      'Cancellation assignment payment', 
                      'Repurchase bill',
                      'Cancel invoice debtor',
                      'Cancel invoice company',
                      '',
                      'Goods dispute bill',
                      'Cancellation reminder invoice','Cancellation of the credit',
                      'Cancellation assignment credit' 
        ]
ttypedict = dict(zip(ttypes, list(range(len(ttypes)))))

if not set(tr.transaction_type.unique()).issubset(set(ttypedict.keys())):
    print("Unexpected transaction types {:}".format(set(tr.transaction_type.unique()).difference(set(ttypedict.keys()))))

tr["ttype"] = tr.transaction_type.apply(lambda x: ttypedict[x])

tr = tr.sort_values(["posting_date", "ttype"], ascending = [True, True])
ttsorted = sorted(ttypedict, key=ttypedict.get) 
ttlist = list(zip(ttsorted, [ttypedict[x] for x in ttsorted]))
for (k,v) in ttlist:
    print("{:} - {:}".format(k, v))

### 3. Checking which columns are instrument attributes and which are trade attributes

The next cell will perform a check on the dataset columns to check which of them are instruments attributes (hence they will contain a unique value for each instrument) and which are trade attributes (having different values for each instrument since they are related to the single trade).

#this is slow
#compiling a dictionary of value lists in each column per instrument
i_g = tr.groupby("uid")
mult_per_i={}
for c in tr.columns:
    #aggregate invoice id and count unique values of c in each grouping:
    ag = i_g.agg({"uid": "first", "invoice_number": "first", c: [lambda x: len(x.unique()), "unique"]})
    exc = ag[c]["unique"][(ag[c]["<lambda>"]>1.)] #instruments with more than one c value
    if len(exc.index)>50: #many instruments with different values
        print(c, ": ", len(exc.index), " instruments")
    elif len(exc.index) == 0: #instrument level column for all trades
        print(c, ": ok")
    else: #some instruments with differenct values
        print(c, ": ", len(exc.index), " instruments", exc)
        
    mult_per_i[c] = exc


### 4. Build instruments structure

#build the instruments structure, preserving all data in trx
#add transaction type enumerator column

#count of entries in mult_per_i
ncoll={}
for k in mult_per_i.keys():
    ncoll[k] = mult_per_i[k].shape[0]

i_g = tr.groupby("uid")
#aggregate by first value...
#dictionary of operation to perform:
agg_cdict = dict(zip(tr.columns, ["first" for c in tr.columns]))
#...unless there are instruments with more than one value for which use list...
for k,v in ncoll.items():
    if v>0: 
        agg_cdict[k]=(lambda x: list(x))
#...unless max makes more sense
agg_cdict["deduction_amount"] = "max"
inst = i_g.agg(agg_cdict)
inst.head()

### 5. Save the instruments dataset

filename = "01_instrumentsdf.pkl"
inst.to_pickle(datafolder + filename)
