==> 02_buyers_sellers_df.ipynb <==
# Buyers and Sellers

This notebook will take the output from step 01 ('01_instrumentsFeatures.ipynb' notebook) and generate separate dataframe of buyers and sellers, in order to perform some data exploration/visualization on the provided portfolio.

## Data import

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import datetime

#Instruments+features dataset
#import data
filename = "02_instrumentsdf_2.pkl"
datafolder = "../data/"

inst = pd.read_pickle(datafolder+filename)
inst.head()

inst.has_impairment1.sum()

#set up buyer dataframe
b_g = inst.groupby("debtor_name_1")
#build an aggregation dictionary

#aggregate by keeping a list
agg_bdict = dict(zip(["customer_id", "customer_name_1", "debtor_id", "uid",
                      "has_impairment1", "is_pastdue90", "is_pastdue180",
                      "has_prosecution", "is_open"
                     ],  ["nunique"]*4 + ["sum"]*6  ))

                                                           
#aggregate by first value...
buyer = b_g.agg(agg_bdict)
buyer.head(4)

#set up seller dataframes
s_g = inst.groupby("customer_name_1")
#build an aggregation dictionary

#aggregate by keeping a list
agg_sdict = dict(zip(["customer_id", "debtor_name_1", "debtor_id", "uid",
                      "has_impairment1", "is_pastdue90", "is_pastdue180",
                      "has_prosecution", "is_open"
                     ],
                   ["nunique"]*4 + ["sum"]*6  ))

                                                           
#aggregate by first value...
seller = s_g.agg(agg_sdict)
seller.head(4)

prefix = "inst_"
seller.to_pickle(datafolder + prefix + "seller.pkl")
buyer.to_pickle(datafolder + prefix + "buyer.pkl")
