==> 05_preproc_pipeline_1.ipynb <==
# Preprocessing pipeline - benchmark models

This notebook starts from the output at step 4 in order to prepare the data to be put into the benchmark models for predictions

import numpy as np
import pandas as pd
import pickle
import datetime
from scripts_ml.preprocessing_pipeline import *

## Importing data and cleaning unnnecessary instruments

datafolder = "../data/"
inputfilename = '03_instrumentsdf_deg1stats.pkl'

#feature selection
feat_str = ['currency']
feat_quant = ['has_purchase', 'dd_value_date', 'cd_lent_c', 'cd_repaid_c', 'cd_impaired1_c', 'cd_pastdue90_c', 'cd_trend_a',  'c_lent_c', 
             'c_repaid_c', 'c_impaired1_c', 'c_pastdue90_c', 'c_trend_a', 'cd_lent_c', 'd_repaid_c', 'd_impaired1_c', 'd_pastdue90_c',
             'd_trend_a', 'd_we_payment_share']
feat_exp = ['invoice_amount', 'purchase_amount']
feat_date = ['invoice_date']

df = pd.read_pickle(datafolder+inputfilename)
df.head()

#drop all instruments that are not due yet, since they can't be labelled
print("{:} instruments that are not due yet, dropping...".format(sum(~df.is_due)))
df=df.loc[df.is_due, :]
print("{:} instruments remaining".format(df.shape[0]))

## Pipeline

targets = ['has_impairment1', 'is_pastdue90', 'is_pastdue180']

### Generating preprocessed data for impairment, pastdue90 and pastdue180 credit events - Shuffle mode

pfixes = ['imp_', 'p90_', 'p180_']
output_path = datafolder+"/preproc_traintest/"

for t in range(len(targets)):
    y_train, X_train, y_test, X_test, feature_labels = preprocessing_pipeline(df, feat_str, feat_quant, feat_exp, feat_date, targets[t], 'invoice_date', 
                                                                              'benchmarks_shuffle', trainsize = .80, testsize = .20,
                                                                         save_to_file=True, outputpath=output_path, prefix=pfixes[t])

### Generating preprocessed data for impairment, pastdue90 and pastdue180 credit events - Timewise mode

for t in range(len(targets)):
    if t!=2:
        tdate = datetime.datetime(2018, 4, 30)
        y_train, X_train, y_test, X_test, feature_labels = preprocessing_pipeline(df, feat_str, feat_quant, feat_exp, feat_date, targets[t],
                                                                             'invoice_date', 'benchmarks_time', timewise=True, testdate = tdate,
                                                                             save_to_file=True, outputfolder=output_path, prefix=pfixes[t])
    else:
        tdate = datetime.datetime(2018, 2, 20)
        y_train, X_train, y_test, X_test, feature_labels = preprocessing_pipeline(df, feat_str, feat_quant, feat_exp, feat_date, targets[t],
                                                                             'invoice_date', 'benchmarks_time', timewise=True, testdate = tdate,
                                                                             save_to_file=True, outputfolder=output_path, prefix=pfixes[t])
