==> 06_models_loop_benchmarks_p180.ipynb <==
# Models loop automation - Pastdue180

This notebook uses as inputs the outputs from the pipeline ('05_preproc_pipeline_1.ipynb' notebook) and and performs model assessment of performances and general exploration for the transaction credit events prediction "is_pastdue180".  
All the experiments produce useful data for visualizing results and can be tracked on MLflow.

import pandas as pd
import numpy as np
import pickle

from scripts_ml.models_utils import *
from scripts_viz.visualization_utils import *

from sklearn.linear_model import SGDClassifier, LogisticRegression
from sklearn.ensemble import RandomForestClassifier

from bokeh.io import show, output_notebook
output_notebook()

preproc_folder = "benchmarks_shuffle"
datafolder = "../data/preproc_traintest/"+preproc_folder+'/'
output_path = "../data/models/"

prefixes_shuffle = ['shuffle_p180_']
postfixes_shuffle = ['_19072_750']

#Linear model Stochastic Gradient Descent
sgd = SGDClassifier(random_state=42, max_iter=250, loss='log', tol=0.0001)

rf = RandomForestClassifier(random_state=42,
                               n_estimators=200,
                               max_leaf_nodes=40,
                               class_weight="balanced",
                               n_jobs=7)

models = [sgd, rf]

## Experiment in Shuffle Mode

expname = preproc_folder+"_p180"

experiment_shuffle = models_loop(models, datafolder, prefixes_shuffle, postfixes_shuffle, mlf_tracking=True, save_model=True,
                                experiment_name=expname, save_results_for_viz=False)

## Experiment in time mode

preproc_folder = "benchmarks_time"
datafolder = "../data/preproc_traintest/"+preproc_folder+'/'
prefixes_time = ['time_2018-02-20_p180_']
postfixes_time = ['_190710_745']

expname = preproc_folder+"_old_p180"

experiment_time_seq = models_loop(models, datafolder, prefixes_time, postfixes_time, 
                              mlf_tracking=True, save_model=True,
                                experiment_name=expname, save_results_for_viz=False)

## Experiment in sequential time mode

#preproc_folder = "benchmarks_time"
#datafolder = "../data/preproc_traintest/"+preproc_folder+'/'
#prefixes_time = ['time_2018-02-20_p180_']
#postfixes_time = ['_190710_745']

#expname = preproc_folder+"_p180"

#experiment_time_seq = models_loop(models, datafolder, prefixes_time, postfixes_time, 
#                              timeSeqValid = True, train_window = 12000, test_window = 3000,
#                              mlf_tracking=True, save_model=True,
#                                experiment_name=expname, save_results_for_viz=False)
