==> 08_models_loop_benchmarks_opt_imp.ipynb <==
# Models loop automation - Impairment tuned models

This notebook uses as inputs the outputs from the pipeline ('05_preproc_pipeline_1.ipynb' notebook) and performs model assessment of performances and general exploration for the transaction credit events prediction "has_impairement1".  
All the experiments produce useful data for visualizing results and can be tracked on MLflow.  

import pandas as pd
import numpy as np
import pickle

from scripts_ml.models_utils import *
from scripts_viz.visualization_utils import *

from sklearn.linear_model import SGDClassifier, LogisticRegression
from sklearn.ensemble import RandomForestClassifier

from bokeh.io import show, output_notebook
output_notebook()

#linear model Stochastic Gradient Descent with semi-optimized hyperparameters using random search

sgd_rs_1 = SGDClassifier(random_state=42, max_iter=80, loss='log', learning_rate='adaptive', eta0=0.001, tol=0.0001) #from_benchmark_imp_time_tuning

sgd_models = [sgd_rs_1]

#random forest models parameters from randomized grid search

#from benchmark_imp_time_tuning
rf_opt_1 = {'n_estimators': 70,
 'min_samples_split': 12,
 'min_samples_leaf': 4,
 'max_leaf_nodes': 90,
 'max_features': 'sqrt',
 'max_depth': 20,
 'bootstrap': True,
 'random_state':42,
 'class_weight':"balanced",
  'n_jobs': 7}

rf_parameters = [rf_opt_1]

rf_models = [RandomForestClassifier(**params) for params in rf_parameters]

models = sgd_models+rf_models

prefixes_shuffle = ['shuffle_imp_']
postfixes_shuffle = ['_19072_750']
preproc_folder = "benchmarks_shuffle" #folder with the correct preprocessing data
expname = preproc_folder+"_opt_imp" #experiment name

datafolder = "../data/preproc_traintest/"+preproc_folder+'/'
output_path = "../data/models/"

## Experiment in Shuffle Mode

#experiment_shuffle = models_loop(models, datafolder, prefixes_shuffle, postfixes_shuffle, mlf_tracking=True, save_model=True,
#                                experiment_name=expname, save_results_for_viz=False)

## Experiment in Time mode

preproc_folder = "benchmarks_time" #folder with the correct preprocessing data
expname = preproc_folder+"_opt_imp" #experiment name
datafolder = "../data/preproc_traintest/"+preproc_folder+'/'
prefixes_time = ['time_2018-04-30_imp_']
postfixes_time = ['_190710_745']

experiment_time = models_loop(models, datafolder, prefixes_time, postfixes_time, 
                              timeSeqValid = True, train_window = 12000, test_window = 3000,
                              mlf_tracking=True, save_model=True,
                                experiment_name=expname, save_results_for_viz=False)
