==> 12_models_loop_enriched_p90.ipynb <==
# Models loop automation - pastdue90

This notebook uses as inputs the outputs from the pipeline 2 ('11_preproc_pipeline_1.ipynb' notebook) and performs model assessment of performances and general exploration for the transaction credit events prediction "pastdue90".  
All the experiments produce useful data for visualizing results and can be tracked on MLflow.

import pandas as pd
import numpy as np
import pickle

from scripts_ml.models_utils import *
from scripts_viz.visualization_utils import *

from sklearn.linear_model import SGDClassifier, LogisticRegression
from sklearn.ensemble import RandomForestClassifier

from bokeh.io import show, output_notebook
output_notebook()

preproc_folder = "enriched_shuffle"
datafolder = "../data/preproc_traintest/"+preproc_folder+'/'
output_path = "../data/models/"

prefixes_shuffle = ['shuffle_p90_bg_']
postfixes_shuffle = ['_190721_1655']

#Linear model Stochastic Gradient Descent
sgd = SGDClassifier(random_state=42, max_iter=250, loss='log', tol=0.0001)

rf = RandomForestClassifier(random_state=42,
                               n_estimators=200,
                               max_leaf_nodes=40,
                               class_weight="balanced",
                               n_jobs=7)

models = [sgd, rf]

## Experiment in Shuffle Mode

expname = preproc_folder+"_p90"

experiment_shuffle = models_loop(models, datafolder, prefixes_shuffle, postfixes_shuffle, mlf_tracking=True, save_model=True,
                                experiment_name=expname, save_results_for_viz=False)

## Experiment in Sequential Time mode preventing time leak

preproc_folder = "enriched_time_seq"
datafolder = "../data/preproc_traintest/"+preproc_folder+'/'
prefixes_time_seq = ['time_2018-04-30_p90_bg_']
valid_code = '_val_26000_2000_'
postfixes_time_seq_val = ['_190824_1722']
postfixes_time_seq = ['_190824_1230']
preproc_folder = "enriched_time_seq"
datafolder = "../data/preproc_traintest/"+preproc_folder+'/'
indexfile = '_fold_indexes'
expname = preproc_folder+valid_code.split('_val_')[1][:-1]+"_p90"

experiment_time_seq = models_loop_time_leak(models, datafolder, prefixes_time_seq, postfixes_time_seq, valid_code, postfixes_time_seq_val,
                                            indexfile, experiment_name = expname, mlf_tracking=True, save_model=True,
                                save_results_for_viz=False)
