==> 14_models_loop_enriched_opt_imp.ipynb <==
# Models loop automation - impairment tuned models

This notebook uses as inputs the outputs from the pipeline 2 and performs model assessment of performances and general exploration for the transaction credit events prediction "impairment".  
All the experiments produce useful data for visualizing results and can be tracked on MLflow.  
The models' hyperparameters used at this stage have been tuned using random grid search in the notebook "13_enriched_models_imp_tuning.ipynb"

import pandas as pd
import numpy as np
import pickle

from scripts_ml.models_utils import *
from scripts_viz.visualization_utils import *

from sklearn.linear_model import SGDClassifier, LogisticRegression
from sklearn.ensemble import RandomForestClassifier

from bokeh.io import show, output_notebook
output_notebook()

#linear model Stochastic Gradient Descent with optimized hyperparameters using random search
sgd_rs = SGDClassifier(random_state=42, max_iter=140, loss='log', learning_rate='optimal', eta0=0.01, tol=0.0001) 

#random forest models parameters from randomized grid search

#first randomized search
rs1 = {'n_estimators': 360,
 'min_samples_split': 15,
 'min_samples_leaf': 1,
 'max_leaf_nodes': 80,
 'max_features': 'sqrt',
 'max_depth': 70,
 'bootstrap': True}


rf_rs_1 = RandomForestClassifier(random_state=42,
                               n_estimators=rs1['n_estimators'],
                               min_samples_split=rs1['min_samples_split'],
                               min_samples_leaf=rs1['min_samples_leaf'],
                               max_features = rs1['max_features'],
                               max_leaf_nodes=rs1['max_leaf_nodes'],
                               max_depth=rs1['max_depth'],
                               bootstrap=rs1['bootstrap'],
                               class_weight="balanced",
                               n_jobs=7)

models = [sgd_rs, rf_rs_1]

#prefixes_shuffle = ['shuffle_imp_bg_']
#postfixes_shuffle = ['_190721_1655']
#preproc_folder = "enriched_shuffle" #folder with the correct preprocessing data
#expname = preproc_folder+"_opt_imp" #experiment name

#datafolder = "../data/preproc_traintest/"+preproc_folder+'/'
#output_path = "../data/models/"

## Experiment in Shuffle Mode

#experiment_shuffle = models_loop(models, datafolder, prefixes_shuffle, postfixes_shuffle, mlf_tracking=True, save_model=True,
#                                experiment_name=expname, save_results_for_viz=True)

## Experiment in Sequential Time mode preventing time leak

preproc_folder = "enriched_time_seq"
datafolder = "../data/preproc_traintest/"+preproc_folder+'/'
prefixes_time_seq = ['time_2018-04-30_imp_bg_']
valid_code = '_val_24000_6000_'
postfixes_time_seq_val = ['_190824_122']
postfixes_time_seq = ['_190824_1138']
preproc_folder = "enriched_time_seq"
datafolder = "../data/preproc_traintest/"+preproc_folder+'/'
indexfile = '_fold_indexes'
expname = preproc_folder+valid_code.split('_val_')[1][:-1]+"_opt_imp"

experiment_time_seq = models_loop_time_leak(models, datafolder, prefixes_time_seq, postfixes_time_seq, valid_code, postfixes_time_seq_val,
                                            indexfile, experiment_name = expname, mlf_tracking=True, save_model=True,
                                save_results_for_viz=False)
